+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * `pilau_get_users_by_capability()`
 * Adjusted `pilau_responsive_image` to work better with retina by default
 * Filter to prevent non-admins from creating admin accounts
+* Filter to prevent non-admins from editing admin accounts
 
 ## 2.1.3 (2015-08-19)
 * Added slug stopwords removal to compensate for Yoast SEO not working